requirements.txt
click==8.1.7
-e git+ssh://git@github.com/oiao/clut.git@df123e4a32a37f1aae66a457bd666961ad1fdc4f#egg=clut
colorama==0.4.6
colormath==3.0.0
contourpy==1.2.1
cycler==0.12.1
docopt==0.6.2
fonttools==4.53.0
imageio==2.34.1
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pylut==1.4.9
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
scipy==1.13.1
six==1.16.0
tifffile==2024.5.22
